# Use escape-html instead of he lib for html encoding for performance